Clean up fixture structure
