+++ tests/test_utils.py
@@ -8,11 +8,11 @@ import pytest
 import xarray as xr
 
 
-def test_utils_smoke(dummy_temp_data, tmp_path):
+def test_utils_smoke(get_dummy_bathymetry_data, tmp_path):
 
     # Test logger
     logger = utils.setup_logger("test_utils")
     assert logger is not None
 
     # Test export_dataset
-    utils.export_dataset(dummy_temp_data, tmp_path / "temp_test.nc")
+    utils.export_dataset(get_dummy_bathymetry_data, tmp_path / "temp_test.nc")
